mithril_era

Module adapters

Module dedicated to EraReaderAdapter implementations.

Structs

EraMarkersPayloadCardanoChain

Era markers payload

EraReaderAdapterBuilder

Era adapter builder

EraReaderBootstrapAdapter

The goal of the bootstrap adapter is to advertise for the first existing Era while it does not exist yet on any backend. This adapter is intended to be removed once Eras are effectively written in a backend.

EraReaderCardanoChainAdapter

Cardano Chain adapter retrieves era markers on chain

EraReaderDummyAdapter

Dummy adapter is intended to be used in a test environment (end to end test) to simulate not yet activated Eras.

EraReaderFileAdapter

File adapter is intended to be used in a test environment to simulate eras transitions.

Enums

EraReaderAdapterType

Type of era reader adapters available